Then we run `rebar3 compile`
When we run so, rebar3 creates a lock file with the specific hashes of
what is to be fetched, to ensure that builds are deterministic.

Imagine the branch or tag name changing the commit they point to and
breaking our system!

The problem is that, when upgrading or changing dependencies, you have
to "unlock it" using `rebar3 unlock dep-name`, and then changing
`rebar3.config` and recompiling. Afterwards, remember to keep both
rebar3.config and rebar3.lock in your git repo!