Add jiffy to our application
To make parsing simple, we can just take advantage of existing protocols
and libraries. I propose JSON, a well-known file format for which
parsers exists in virtually all languages, so that all kinds of clients
can easily build their own parsers. To parse JSON in erlang, the
de-facto standard is `jiffy` (https://github.com/davisp/jiffy), a set of
NIFs to encode and decode JSON.

What is a NIF? Native Implemented Function, is a function implemented in
a native (compiled) language, usually C, that binds to an erlang caller.
This is often done for performance reasons, and it's usually ideal for
those things that are done extremely often and are essentially
algorithmic and non-parallelisable, like, parsing JSON.

It is a dangerous business though, be very careful, because these nifs
are ran on the same environment than the whole Erlang VM, and as so,
they have access to everything the VM has. At best, a crash on your nif
will crash the _entire_ Erlang VM. At worst, you could be introducing
nasty security bugs or breaking the latency properties of the VM.

If you're curious, here's some documentation:
http://erlang.org/doc/tutorial/nif.html