Parser template
We need a communication protocol, a way for client and server to agree
on an algorithmically parseable binary format for the messages they
exchange over the network. This formatting protocol, to call it somehow,
is independent of the distribution of such messages, and as such, it can
be implemented and tested separately, and used where necessary.
Theoretically, this could be on a different library, one of those
`rebar3.config` dependencies ;)