Add `proper`'s include file to the test suite
We add it through the `include_lib` directive. This has two important
things to be said:

One is, what is that `include_lib` directive? In erlang, there are two
methods to include information that has been stated in other places. One
is through the `-include` directive, to which you give a relative path
to the file you're currently editing. The other one is the
`include_lib`, which will find the path on the libraries that come with
the installation of your OTP release. In this case, `rebar3` is managing
this installation and all of its libraries, through the `deps` section
of the `rebar.config` file.

The second is about proper. This include file is the one that has
definitions for the macros and types that proper will use, and it might
also contain definitions for other more powerful compile-time
transformations: if you feel adventurous, look up for "parse
transforms" :)