Generate correct examples
In this property, there are two edges cases: when we basically send
nothing, and when we sent the EOT character. The property behaves
differently for those, hence, we don't want to test the property for
them, because we know that it won't hold.

So what we do is that we modify the generator, to tell him to skip these
two values, using
```
?SUCHTHAT(Msg, binary(), Msg =/= <<>> andalso Msg =/= <<4>>)
```
Which reads as
"Such that `Msg` is a binary not equal to `<<>>` and also not equal to <<4>>"