Refactor echo tests to a group on its own
As we're going to add more tests related to this same protocol, we can
group them together. Now we see that functionality like starting ranch
is a functionality of the whole suite, and that of starting the echo
listener is something that relates to the whole echo_protocol group, so
we move logic to the appropriate places.