Implement `parser:encode/1`
Voilà! If our message structure is nothing more than a map,
`jiffy:encode/2` already turns that into a valid JSON in binary that we
can use to send. Very simple!

If what we're passing was not to have a proper Erlang-like JSON
structure, jiffy's API says it would throw a `{invalid_ejson, _}`
as an exception, so we can catch it and return the domain-specific atom
we want.