Prototype `parser:decode/1`
Decoding a message would be a similar business than encoding. Thanks to
choosing JSON as a text protocol, parsing a binary into an Erlang-like
JSON structure comes for free using JSON, so getting a working map is
done for us.

What we need to implement is our particular business logic. What fields
are mandatory for us? How should each type of stanza look like? It is
all that logic that we need to implement. Here there are two
placeholders: "message" for simply messaging somebody, and "status", for
giving information about errors, successful logins, server informing
about other clients, etc.

For these two, we have two placeholder for checks, `parse_message` and
`parse status`, that need to be implemented.