Template new test for the real messenger protocol
This is again a placeholder. We're slightly defining the interface for
our implementation already. Notice for example the call to
`ranch:start_listener`: we already named the callback module,
`messenger_protocol`, and the port it will listen to, `5556`.
Now we only need to define, in the test, how does a successful login
looks like.