Test a successful login
And here it is, at last, a very simple test, to check what we've been
expecting all along!

We can expect that many more tests will do log in for diverse clients,
so we can already take a helper `do_login` function aside, with the
appropriate parameters: the nickname, and the expected answer.