Refactor testsuite
As we plan now to test a scenario where we have two users instead of
one, we might foresee that we'll as well have scenarios with three,
four, five, or any amount or clients.

So to keep code organised, a smart move would be to group tests
according to the amount of clients this test will require, hence the
whole initialisation logic remains generic in the corresponding
`init_per_group` and `init_per_testcase`.

In the `init_per_group`, we encode the amount of clients this group
requires. Then on each `init_per_testcase`, we generate this amount of
users, with the help of a helper function that generates the expected
amount of clients, and we add them all to the `Config`, so that they can
be used in the test. In the `end_per_testcase`, we extract this list of
users, and close their connections.

This helper function generates tuples tagged with an atom `user` and the
user's number:
```
     binary_to_atom(<<"user", (integer_to_binary(X))/binary>>, utf8)
```

And we add this as the tag to a tuple that contains as a value, the
given socket and nickname for this user. If you check the code, you'll
see that I've coded the testcase's name into the user's nickname, so you
might wonder why do I do something like that? By experience, this is
good for debugging. When tests fails, the information of this test will
be in the logs of both your testsuite and your running application, so
being able to distinguish testcases by the logged nickname will be
useful ;)

Why does it all look like a `{tag, {Socket, Nickname}}`? Well,
_commontest_'s `?config` macro will precisely extract the first tagged
tuple that it finds in the config, given a tag.