Test message to non-existent user
What happens when a recipient doesn't exist? Typical behaviour for a
messaging service is to store it offline until the recipient comes, but
we need to actually think about the reasonale behind this decision.
Basically, it is a business decision, where we weight user experience by
server complexity. Who pays the price? It is often decided that the
server does so, as, the simpler the user experience is, the more likely
it is to get a larger client base, which will pay off server costs. But
this makes the server complex. For now, we can keep it simple for us: we
tell the user we cannot deliver this message, and we forget about it.
And it is now the client's responsibility (his app can do it
automatically) to take on the retry logic.

Also refactor the tests to avoid duplication of the message delivery
logic, as now suddenly sending a login stanza and a regular message seem
very similar.