A ranch messenger protocol
Having in place some test, we know a lot about how the messenger
protocol should look like, and using our previous `basic_protocol` as a
template, it's straightforward to get started.

There is just one detail to distinguish: stages. We first log in, choose
a nickname —and be sure it is available!—, and only then we start
chatting. This can easily be distinguished by pattern matching on the
loop function head.