Create a session table
Now, a very important detail: we need to keep track of the pairing
between a nickname, and another user's client-to-server process. Note
that ranch initialises a new process on every client request, and it is
this process what communicates with the client. So to deliver a message
to a client, given his nickname, we only need this process. And then we
can send the erlang way!