Implement inserting a session
Note very importantly how we extract the `Pid`: it is, exclusively, the
calling process. To prevent errors, it is a good design that a process
can only register himself. Otherwise, this `insert_session` function
could be used in wild ways, registering non-existent processes or worse,
letting processes steal nicknames from each other!

The next thing we need to make sure, is that we're not overriding
someone else's nickname. That is what the function `ets:insert_new/2`
does, it inserts the value only if it is new, and returns as a boolean
whether the value was actually inserted. If it was not, it is because
the nickname was already in use, hence returning
`{error, nickname_conflict}`