Implement removing a session
When we remove a session, we query the ets table by the Nickname. The
first check that is important to do is to be sure that the stored
session matches perfectly what is being requested to remove: that is,
that both `Nickname` and the calling `Pid` match. If so, we delete this
session from the table.

Other possibilities are that the calling `Pid` doesn't match the stored
one, and that could mean that somebody is requesting removing a session
that is not his.