Implement getting a session pid by the nickname
Now, to finally know to whom to send a message, we need his pid, so
given a nickname, we need to get the paired pid. This is a simple
getter, and the logic is very similar than that of removing a pid: by
just doing a `ets:lookup/2`, and pattern-matching on the result value.